--- src/app/(headerless)/login/hooks/useLoginPage.ts
@@ -1,11 +1,0 @@
-import { useIsLogIn } from '@/hooks/useIsLogIn';
-import { useRouter } from 'next/navigation';
-import { useEffect } from 'react';
-
-export default function useLoginPage() {
-  const { isLogin } = useIsLogIn();
-  const router = useRouter();
-  useEffect(() => {}, [isLogin, router]);
-
-  return { isLogin };
-}
